Add VNPay Payment

+++ src/main/java/controller/user/OrderServlet.java
@@ -1,6 +1,5 @@
 package controller.user;
 
-import entity.CartItem;
 import entity.Address;
 import entity.Product;
 import entity.ShopOrder;
@@ -28,12 +27,9 @@ public class OrderServlet extends HttpServlet {
         Address tempAddress = (Address) request.getSession().getAttribute("temporaryAddress");
         if (tempAddress != null) {
             String PayType = String.valueOf(request.getSession().getAttribute("PayType"));
-            System.out.println(PayType);
             if (PayType.equals("null")) {
-                request.setAttribute("PayMessError", "You must chooser payment method");
+                request.setAttribute("PayMessError", "You must choose a payment method");
                 request.getRequestDispatcher("checkout.jsp").forward(request, response);
-            } else if (PayType.equals("VNPAY")) {
-                request.getRequestDispatcher("ajaxServlet").forward(request, response);
             } else {
                 int userID = Integer.parseInt(request.getParameter("UserID"));
                 CartItemDAO cartItemDAO = new CartItemDAO();
@@ -41,41 +37,46 @@ public class OrderServlet extends HttpServlet {
                 OrderDetailDAO orderDetailDAO = new OrderDetailDAO();
                 AddressDAO ad = new AddressDAO();
                 List<Product> cartItems = cartItemDAO.getUserItem(userID);
-
-                // Tạo đơn hàng
                 ShopOrderDAO shopOrderDAO = new ShopOrderDAO();
                 ShopOrder shopOrder = shopOrderDAO.getLatestOrder();
+                ShopOrder shopOrderReturn = (ShopOrder) session.getAttribute("shopOrder");
+                if (shopOrderReturn == null) {
+                    response.sendRedirect("cart");
+                    return;
+                }
                 int orderID = shopOrder.getShop_orderID();
-
-                // Đặt hàng cho từng sản phẩm trong giỏ hàng
+                double totalOrderAmount = 0;
                 for (Product cartItem : cartItems) {
                     String productID = String.valueOf(cartItem.getProductID());
                     String variationID = String.valueOf(cartItem.getVariationID());
-                    int quantity = cartItem.getQty_in_cart();  // Lấy số lượng từ giỏ hàng
-                    double price = cartItem.getPrice() * quantity;  // Tính toán giá trị dựa trên số lượng
-
-                    // Thực hiện đặt hàng cho sản phẩm
+                    int quantity = cartItem.getQty_in_cart();
+                    double price = cartItem.getPrice() * quantity;
+                    totalOrderAmount += price;
                     orderDetailDAO.insert(orderID, productID, variationID, quantity, price);
                     VariationDAO variationDAO = new VariationDAO();
-
-                    //Giảm quanty in stock
                     Variation var = variationDAO.getVariation(productID, cartItem.getColor_Name(), cartItem.getSize_Name());
                     variationDAO.updateVariation(cartItem.getProductID(), var.getColor_ID(), var.getSize_ID(), var.getQtu_in_stock() - cartItem.getQty_in_cart(), var.getProduct_img_ID(), var.getVariationID());
-
-                    // Xóa sản phẩm đã đặt hàng khỏi giỏ hàng
-                    cartItemDAO.deleteCartItemByProdID(productID, variationID);
-                    pd.reduceQuantityOfProduct(productID, variationID, quantity);
-                    ad.setAddressIDtoShopOrder();
-                    shopOrderDAO.setOrderTotal();
+                    if (!PayType.equals("VNPAY")) {
+                        cartItemDAO.deleteCartItemByProdID(productID, variationID);
+                        pd.reduceQuantityOfProduct(productID, variationID, quantity);
+                    }
                 }
-                System.out.println(orderID);
+                shopOrderDAO.setOrderTotal();
+                ad.setAddressIDtoShopOrder();
                 int saleId = shopOrderDAO.getSaleWithMinOrder();
-                System.out.println(saleId);
                 shopOrderDAO.insertSaleAssignment(orderID, saleId);
-                session.removeAttribute("temporaryAddress");
-                session.removeAttribute("PayType");
-                request.setAttribute("OrderMessage", "Order Placed");
-                request.getRequestDispatcher("thanks.jsp").forward(request, response);
+                if (PayType.equals("VNPAY")) {
+                    session.setAttribute("orderId", orderID);
+                    long paymentPrice = (long) Math.round(totalOrderAmount + 36000);
+                    String paymentUrl = "vnpay" + "?amount=" + paymentPrice;
+                    response.sendRedirect(paymentUrl);
+                } else {
+                    session.removeAttribute("temporaryAddress");
+                    session.removeAttribute("PayType");
+                    request.setAttribute("OrderMessage", "Order Placed");
+                    request.setAttribute("TotalOrderAmount", totalOrderAmount);
+                    request.getRequestDispatcher("thanks.jsp").forward(request, response);
+                }
             }
         } else {
             request.setAttribute("OrderMessageError", "You must enter shipping information");
